requirements.txt
kafka-python
requests
python-dotenv
sentence-transformers
qdrant-client
tqdm
pytz
sentencepiece
fastapi
uvicorn